pyproject.toml
[project]
name = "orderservice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.3",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "dependency-injector>=4.48.3",
    "fastapi>=0.135.1",
    "pydantic-settings>=2.13.1",
    "python-json-logger>=4.0.0",
    "ruff>=0.15.6",
    "sqlalchemy>=2.0.48",
    "tenacity>=9.1.4",
    "uvicorn>=0.41.0",
]
